Support `export * as Foo from 'foo';`
Fix #226.
