+++ ecmascript/ast/src/lib.rs
@@ -35,9 +35,9 @@ pub use self::{
     lit::{Bool, Lit, Null, Number, Regex, RegexFlags, Str},
     module::{Module, ModuleItem, Script},
     module_decl::{
-        ExportAll, ExportAllAs, ExportDefaultDecl, ExportSpecifier, ImportDecl, ImportDefault,
-        ImportSpecific, ImportSpecifier, ImportStarAs, ModuleDecl, NamedExport,
-        NamedExportSpecifier,
+        ExportAll, ExportDefaultDecl, ExportSpecifier, ImportDecl, ImportDefault, ImportSpecific,
+        ImportSpecifier, ImportStarAs, ModuleDecl, NamedExport, NamedExportSpecifier,
+        NamespaceExportSpecifier,
     },
     operators::{AssignOp, BinaryOp, UnaryOp, UpdateOp},
     pat::{
