+++ ecmascript/ast/src/module_decl.rs
@@ -17,7 +17,6 @@ pub enum ModuleDecl {
 
     ExportDefaultExpr(Box<Expr>),
     ExportAll(ExportAll),
-    ExportAllAs(ExportAllAs),
     TsImportEquals(TsImportEqualsDecl),
     TsExportAssignment(TsExportAssignment),
     TsNamespaceExport(TsNamespaceExportDecl),
@@ -35,14 +34,6 @@ pub struct ImportDecl {
 #[ast_node]
 pub struct ExportAll {
     pub span: Span,
-    pub src: Str,
-}
-
-/// `export * as Foo from 'mod'`
-#[ast_node]
-pub struct ExportAllAs {
-    pub span: Span,
-    pub name: Ident,
     pub src: Str,
 }
 
@@ -96,8 +87,16 @@ pub struct ImportSpecific {
 
 #[ast_node]
 pub enum ExportSpecifier {
+    Namespace(NamespaceExportSpecifier),
     Default(Ident),
     Named(NamedExportSpecifier),
+}
+
+/// `export * as foo from 'src';`
+#[ast_node]
+pub struct NamespaceExportSpecifier {
+    pub span: Span,
+    pub name: Ident,
 }
 
 #[ast_node]
