+++ ecmascript/transforms/src/modules/amd/mod.rs
@@ -64,9 +64,6 @@ impl Fold<Module> for Amd {
             };
 
             match decl {
-                ModuleDecl::ExportAllAs(..) => {
-                    unreachable!("export * as Foo from 'foo'; should be removed by previous pass")
-                }
                 ModuleDecl::Import(import) => self.scope.insert_import(import),
 
                 ModuleDecl::ExportAll(..)
@@ -252,9 +249,13 @@ impl Fold<Module> for Amd {
                             for NamedExportSpecifier { orig, exported, .. } in
                                 export.specifiers.into_iter().map(|e| match e {
                                     ExportSpecifier::Named(e) => e,
-                                    _ => unreachable!(
+                                    ExportSpecifier::Default(..) => unreachable!(
                                         "export default from 'foo'; should be removed by previous \
                                          pass"
+                                    ),
+                                    ExportSpecifier::Namespace(..) => unreachable!(
+                                        "export * as Foo from 'foo'; should be removed by \
+                                         previous pass"
                                     ),
                                 })
                             {
