+++ ecmascript/transforms/src/proposals/export/tests.rs
@@ -1,9 +1,15 @@
-use super::export_namespace_from;
+use super::export;
 use ast::*;
 use swc_common::Fold;
 use swc_ecma_parser::{EsConfig, Syntax};
 
-fn syntax() -> Syntax {
+fn syntax_default() -> Syntax {
+    Syntax::Es(EsConfig {
+        export_default_from: true,
+        ..Default::default()
+    })
+}
+fn syntax_namespace() -> Syntax {
     Syntax::Es(EsConfig {
         export_namespace_from: true,
         ..Default::default()
@@ -11,11 +17,34 @@ fn syntax() -> Syntax {
 }
 
 fn tr() -> impl Fold<Module> {
-    export_namespace_from()
+    export()
 }
 
 test!(
-    syntax(),
+    syntax_default(),
+    |_| tr(),
+    default_es6,
+    r#"export foo from "bar";"#,
+    r#"
+import _foo from "bar";
+export { _foo as foo };
+"#
+);
+
+test!(
+    syntax_default(),
+    |_| tr(),
+    default_compounded_es6,
+    r#"export v, { x, y as w } from "mod";"#,
+    r#"
+import _v from "mod";
+export { _v as v };
+export { x, y as w } from "mod";
+"#
+);
+
+test!(
+    syntax_namespace(),
     |_| tr(),
     namespace_compound_es6,
     r"export * as foo, { bar } from 'bar';",
@@ -26,7 +55,7 @@ export { bar } from 'bar';
 );
 
 test!(
-    syntax(),
+    syntax_namespace(),
     |_| tr(),
     namespace_default,
     "export * as default from 'foo';",
@@ -35,7 +64,7 @@ export { _default as default };"
 );
 
 test!(
-    syntax(),
+    syntax_namespace(),
     |_| tr(),
     namespace_es6,
     "export * as foo from 'bar';",
