+++ ecmascript/transforms/src/proposals/mod.rs
@@ -1,9 +1,5 @@
-pub use self::{
-    class_properties::class_properties, decorators::decorators,
-    export_default_from::export_default_from, export_ns_from::export_namespace_from,
-};
+pub use self::{class_properties::class_properties, decorators::decorators, export::export};
 
 mod class_properties;
 mod decorators;
-mod export_default_from;
-mod export_ns_from;
+mod export;
